+++ src/ethernet_log.c
@@ -1,3 +1,9 @@
+/**
+ * @file ethernet_log.c
+ * @author Aryan Chopra
+ * @brief Logs the ethernet packet in a log file.
+ */
+
 #include <errno.h>
 #include <fcntl.h>
 #include <inttypes.h>
@@ -11,7 +17,21 @@
 
 #define SIZE 100
 
+/**
+ * @brief contains the file descriptor of the log file opened.
+ */
+
 int ethLogFile;
+
+/**
+ * @brief Opens the log file.
+ *
+ *
+ * Opens the log file to log the ethernet headers, in append mode.
+ * Prints an error to the consols and exits the process in case of an error.
+ *
+ * @pre Log file exists at the location.
+ */
 
 void openEthernetLog() {
   ethLogFile = open(LOG_LOCATION"ethernet_log.txt", O_WRONLY | O_APPEND);
@@ -21,9 +41,26 @@ void openEthernetLog() {
   }
 }
 
+/**
+ * @brief Writes the provided character array to the log.
+ *
+ *
+ * @pre File descriptor provided is valid.
+ */
+
 void writeEthLog(char *text) {
   write(ethLogFile, text, strlen(text));
 }
+
+/**
+ * @brief Logs mac address to the log file
+ *
+ *
+ * The function is static to prevent it's symbol being exported and clashin with similar function from other files.
+ * Converts the given address to the commonly legible format and logs it to a file.
+ *
+ * @param[in] mac[6] a character array containing the mac address.
+ */
 
 static void logMacAddress(unsigned char mac[6]) {
   unsigned char logText[4];
@@ -38,7 +75,19 @@ static void logMacAddress(unsigned char mac[6]) {
   writeEthLog(logText);
 }
 
-void logEthernetHeader(EthernetHeader *ethHeader, int incoming) {
+/**
+ * @brief Logs the provided ethernet header.
+ *
+ *
+ * Parses the provided ethernet header and logs the various fields.
+ * Marks the packet as incoming or outgoing based on the flag provided
+ * Calls the appropriate functions to format and log MAC address in legible format.
+ *
+ * @param[in] ethHeader Struct formated with apt names and sizes for various fields of an ethernet header.
+ * @param[in] incoming Flag that marks the packet as incoming or outgoing
+ */
+
+void logEthernetHeader(EthernetHeader *ethHeader, uint8_t incoming) {
   char *text = calloc(SIZE, 1);
 
   if (incoming) {
